implement ErrorDialog for user interactions; IO services now throw exceptions when something is wrong;

+++ src/main/java/com/github/argon/sos/moreoptions/phase/Updateable.java
@@ -1,5 +1,5 @@
 package com.github.argon.sos.moreoptions.phase;
 
 public interface Updateable {
-    void update(float seconds);
+    void update(double seconds);
 }
